% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCIApathfinder.R
\name{get_collection_names}
\alias{get_collection_names}
\title{Get the names of all TCIA collections}
\usage{
get_collection_names()
}
\value{
List containing elements:
\itemize{
  \item \code{collection_names}: character vector of TCIA collection names
  \item \code{content}: parsed API response content
  \item \code{response}: API response
}
}
\description{
Get the names of all TCIA collections
}
\examples{
\dontrun{
get_collection_names()
}

}
\seealso{
\href{https://wiki.cancerimagingarchive.net/display/Public/TCIA+Programmatic+Interface+\%28REST+API\%29+Usage+Guide}{TCIA REST API Usage Guide},
\href{https://wiki.cancerimagingarchive.net/display/Public/TCIA+API+Return+Values}{TCIA API object definitions}
}
