% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCIApathfinder.R
\name{get_new_studies_in_collection}
\alias{get_new_studies_in_collection}
\title{Get studies that have been added to a collection and optionally to a patient since a specified date}
\usage{
get_new_studies_in_collection(collection, date, patient_id = NULL)
}
\arguments{
\item{collection}{TCIA collection name. To get a list of available collection
names, call \code{get_collection_names()}.}

\item{date}{Date in format YYYY-MM-DD}

\item{patient_id}{Patient ID. To get a list of available patient IDs, call \code{get_patients()}.}
}
\value{
List containing elements:
\itemize{
  \item \code{studies}: Data frame of collection, patient ID, and study instance UID
  \item \code{content}: parsed API response content
  \item \code{response}: API response
}
}
\description{
Get studies that have been added to a collection and optionally to a patient since a specified date
}
\examples{
get_new_studies_in_collection("TCGA-BRCA", "2014-01-01")
get_new_studies_in_collection("TCGA-BRCA", "2014-01-01", "TCGA-OL-A66O")

}
