% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCIApathfinder.R
\name{get_patient_ids}
\alias{get_patient_ids}
\title{Get patient IDs}
\usage{
get_patient_ids(collection = NULL)
}
\arguments{
\item{collection}{TCIA collection name. If \code{collection} is \code{NULL},
patients from all collections will be returned. To get a list of available collection
names, call \code{get_collection_names()}.}
}
\value{
List containing elements:
\itemize{
  \item \code{patient_ids}: character vector of patient IDs
  \item \code{content}: parsed API response content
  \item \code{response}: API response
}
}
\description{
Get patient IDs
}
\examples{
get_patient_ids()
get_patient_ids("TCGA-BRCA")

}
