% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCIApathfinder.R
\name{get_patient_studies}
\alias{get_patient_studies}
\title{Get patient study information}
\usage{
get_patient_studies(collection = NULL, patient_id = NULL,
  study_instance_uid = NULL)
}
\arguments{
\item{collection}{TCIA collection name. To get a list of available collection
names, call \code{get_collection_names()}. If \code{collection} is \code{NULL}, information
for all relevant collections will be returned.}

\item{patient_id}{Patient ID. To get a list of available patient IDs, call \code{get_patients()}.
If \code{patient_id} is \code{NULL}, information for all relevant patients will be returned.}

\item{study_instance_uid}{Study instance UID. If \code{study_instance_uid} is \code{NULL}, information
for all relevant study instance UIDs will be returned.}
}
\value{
List containing elements:
\itemize{
  \item \code{patient_studies}: Data frame with columns representing the contents of a PatientStudy object
  as described in \href{https://wiki.cancerimagingarchive.net/display/Public/TCIA+API+Return+Values}{TCIA API Return Values}
  \item \code{content}: parsed API response content
  \item \code{response}: API response
}
}
\description{
Get patient study information
}
\examples{
get_patient_studies()
get_patient_studies(collection = "TCGA-BRCA")
get_patient_studies(patient_id = "TCGA-OL-A6VO")
get_patient_studies(patient_id = "TCGA-OL-A5DA", study_instance_uid = "1.3.6.1.4.1.14519.5.2.1.5382.4002.104582989590517557856962159716")

}
