% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCIApathfinder.R
\name{get_patients}
\alias{get_patients}
\title{Get patient information}
\usage{
get_patients(collection = NULL)
}
\arguments{
\item{collection}{TCIA collection name. If \code{collection} is \code{NULL},
patients from all collections will be returned. To get a list of available collection
names, call \code{get_collection_names()}.}
}
\value{
List containing elements:
\itemize{
  \item \code{patients}: Data frame of patient ID, name, sex, ethnic group, and collection name
  \item \code{content}: parsed API response content
  \item \code{response}: API response
}
}
\description{
Get patient information
}
\examples{
get_patients()
get_patients("TCGA-BRCA")

}
