% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCIApathfinder.R
\name{get_patients_by_modality}
\alias{get_patients_by_modality}
\title{Get patient IDs given a collection name and modality}
\usage{
get_patients_by_modality(collection, modality)
}
\arguments{
\item{collection}{TCIA collection name. To get a list of available collection
names, call \code{get_collection_names()}.}

\item{modality}{Modality name. To get a list of available modality names, call
\code{get_modality_names()}.}
}
\value{
List containing elements:
\itemize{
  \item \code{patient_ids}: Patient IDs
  \item \code{content}: parsed API response content
  \item \code{response}: API response
}
}
\description{
Get patient IDs given a collection name and modality
}
\examples{
get_patients_by_modality("TCGA-BRCA","MR")

}
