% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCIApathfinder.R
\name{get_studies_in_collection}
\alias{get_studies_in_collection}
\title{Get studies in a collection and optionally for a specific patient}
\usage{
get_studies_in_collection(collection, patient_id = NULL)
}
\arguments{
\item{collection}{TCIA collection name. To get a list of available collection
names, call \code{get_collection_names()}.}

\item{patient_id}{Patient ID. To get a list of available patient IDs, call \code{get_patients()}.}
}
\value{
List containing elements:
\itemize{
  \item \code{studies}: Data frame of collection, patient ID, and study instance UID
  \item \code{content}: parsed API response content
  \item \code{response}: API response
}
}
\description{
Get studies in a collection and optionally for a specific patient
}
\examples{
get_studies_in_collection("TCGA-BRCA")
get_studies_in_collection("TCGA-BRCA", "TCGA-OL-A66O")

}
