% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tcia.R
\name{read_tcia}
\alias{read_tcia}
\title{Read TCIA file}
\usage{
read_tcia(file)
}
\arguments{
\item{file}{path to file from TCIA, usually with extension \code{tcia}}
}
\value{
A list of values
}
\description{
Read TCIA file
}
\examples{
file = system.file("doiJNLP-T5OW0OWM.tcia", package = "TCIApathfinder")
res = read_tcia(file)
testthat::expect_equal(res$noOfrRetry, "4")
testthat::expect_equal(length(res$ListOfSeriesToDownload), 82L)
}
